+++ src/components/UI/Account/features/DeleteAccountDialog/DeleteAccountDialog.tsx
@@ -6,7 +6,7 @@ import {
 import { useNotification } from '../../../../../hooks/useNotification';
 import { ERROR_MESSAGE_GENERAL } from '../../../../../constants';
 import { SystemStateEnum } from '../../../../../enums';
-import { DeleteAccountDialogProps } from './interface';
+import { DeleteAccountDialogProps } from './DeleteAccountDialog.interface';
 import { AppIcon } from '../../../Icons';
 import { DeleteAccountMutationProps } from '../../../../../redux/slices/Accounts/interface';
 import { useAppSelector } from '../../../../../redux/hooks';
