+++ src/components/UI/Account/features/SelectAccountDialog/SelectAccountDialog.tsx
@@ -3,7 +3,7 @@ import {
 } from '@mui/material';
 
 import { AccountUI } from '../../Account.interface';
-import { AccountDialogProps } from './interface';
+import { AccountDialogProps } from './SelectAccountDialog.interface';
 import { useAppDispatch, useAppSelector } from '../../../../../redux/hooks';
 import { updateSelectedAccount, updateAccounts } from '../../../../../redux/slices/Accounts/accounts.slice';
 import { ListAccountSelected, ListItemButtonContainer } from './SelectAccountDialog.styled';
